pyproject.toml
[project]
name = "googlemaptrafficconditioncollect"
version = "0.1.0"
description = "NYC Google Maps traffic condition scraper web app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "playwright>=1.44.0",
    "Pillow>=10.3.0",
    "shapely>=2.0.5",
    "apscheduler>=3.10.4",
]
